Add explicit conversions in places where implicit ones are no longer allowed in Solidity 0.8

+++ contracts/UniswapV3Pool.sol
@@ -468,7 +468,7 @@ contract UniswapV3Pool is IUniswapV3Pool, NoDelegateCall {
                     owner: recipient,
                     tickLower: tickLower,
                     tickUpper: tickUpper,
-                    liquidityDelta: int256(amount).toInt128()
+                    liquidityDelta: int256(uint256(amount)).toInt128()
                 })
             );
 
@@ -525,7 +525,7 @@ contract UniswapV3Pool is IUniswapV3Pool, NoDelegateCall {
                     owner: msg.sender,
                     tickLower: tickLower,
                     tickUpper: tickUpper,
-                    liquidityDelta: -int256(amount).toInt128()
+                    liquidityDelta: -int256(uint256(amount)).toInt128()
                 })
             );
